+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/java/JavaTestCompiler.kt
@@ -1,55 +1,25 @@
 package org.jetbrains.research.testspark.core.test.java
 
 import io.github.oshai.kotlinlogging.KotlinLogging
-import org.jetbrains.research.testspark.core.test.TestCasesCompilationResult
 import org.jetbrains.research.testspark.core.test.TestCompiler
-import org.jetbrains.research.testspark.core.test.data.TestCaseGeneratedByLLM
 import org.jetbrains.research.testspark.core.utils.CommandLineRunner
 import org.jetbrains.research.testspark.core.utils.DataFilesUtil
 import java.io.File
 
-
-/**
- * TestCompiler is a class that is responsible for compiling generated test cases using the proper javac.
- * It provides methods for compiling test cases and code files.
- */
-open class JavaTestCompiler(
+class JavaTestCompiler(
     private val javaHomeDirectoryPath: String,
     private val libPaths: List<String>,
     private val junitLibPaths: List<String>,
 ) : TestCompiler {
     private val log = KotlinLogging.logger { this::class.java }
 
-    /**
-     * Compiles the generated files with test cases using the proper javac.
-     *
-     * @return true if all the provided test cases are successfully compiled,
-     *         otherwise returns false.
-     */
-    override fun compileTestCases(
-        generatedTestCasesPaths: List<String>,
-        buildPath: String,
-        testCases: MutableList<TestCaseGeneratedByLLM>,
-    ): TestCasesCompilationResult {
-        var allTestCasesCompilable = true
-        val compilableTestCases: MutableSet<TestCaseGeneratedByLLM> = mutableSetOf()
-
-        for (index in generatedTestCasesPaths.indices) {
-            val compilable = compileCode(generatedTestCasesPaths[index], buildPath).first
-            allTestCasesCompilable = allTestCasesCompilable && compilable
-            if (compilable) {
-                compilableTestCases.add(testCases[index])
-            }
-        }
-
-        return TestCasesCompilationResult(allTestCasesCompilable, compilableTestCases)
-    }
 
     override fun compileCode(path: String, projectBuildPath: String): Pair<Boolean, String> {
         // find the proper javac
         val javaCompile = File(javaHomeDirectoryPath).walk()
             .filter {
-                val isCompilerName = if (DataFilesUtil.isWindows()) it.name.equals("javac.exe") else it.name.equals("javac")
+                val isCompilerName =
+                    if (DataFilesUtil.isWindows()) it.name.equals("javac.exe") else it.name.equals("javac")
                 isCompilerName && it.isFile
             }
             .firstOrNull()
