+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/kotlin/KotlinTestCompiler.kt
@@ -1,39 +1,17 @@
 package org.jetbrains.research.testspark.core.test.kotlin
 
 import io.github.oshai.kotlinlogging.KotlinLogging
-import org.jetbrains.research.testspark.core.test.TestCasesCompilationResult
 import org.jetbrains.research.testspark.core.test.TestCompiler
-import org.jetbrains.research.testspark.core.test.data.TestCaseGeneratedByLLM
 import org.jetbrains.research.testspark.core.utils.CommandLineRunner
 import org.jetbrains.research.testspark.core.utils.DataFilesUtil
 import java.io.File
 
 class KotlinTestCompiler(
-    private val kotlinHomeDirectoryPath: String,
     private val libPaths: List<String>,
     private val junitLibPaths: List<String>,
 ) : TestCompiler {
 
     private val log = KotlinLogging.logger { this::class.java }
-
-    override fun compileTestCases(
-        generatedTestCasesPaths: List<String>,
-        buildPath: String,
-        testCases: MutableList<TestCaseGeneratedByLLM>
-    ): TestCasesCompilationResult {
-        var allTestCasesCompilable = true
-        val compilableTestCases: MutableSet<TestCaseGeneratedByLLM> = mutableSetOf()
-
-        for (index in generatedTestCasesPaths.indices) {
-            val compilable = compileCode(generatedTestCasesPaths[index], buildPath).first
-            allTestCasesCompilable = allTestCasesCompilable && compilable
-            if (compilable) {
-                compilableTestCases.add(testCases[index])
-            }
-        }
-
-        return TestCasesCompilationResult(allTestCasesCompilable, compilableTestCases)
-    }
 
     override fun compileCode(path: String, projectBuildPath: String): Pair<Boolean, String> {
