+++ kotlin/src/main/kotlin/org/jetbrains/research/testspark/kotlin/KotlinBuilderHelper.kt
@@ -1,0 +1,7 @@
+package org.jetbrains.research.testspark.kotlin
+
+class KotlinBuilderHelper {
+    companion object {
+
+    }
+}
